requirements.txt
aniso8601==9.0.1
annotated-types==0.5.0
anyio==3.7.1
attrs==23.1.0
bcrypt==4.0.1
blinker==1.6.2
blis==0.7.10
catalogue==2.0.9
certifi==2023.7.22
cffi==1.15.1
chardet==4.0.0
click==8.1.4
confection==0.1.3
cryptography==41.0.6
cymem==2.0.7
dataclasses==0.6
deprecation==2.1.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0-py3-none-any.whl#sha256=83276fc78a70045627144786b52e1f2728ad5e29e5e43916ec37ea9c26a11212
exceptiongroup==1.1.3
Flask==2.3.2
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
flask-restx==1.3.0
Flask-SQLAlchemy~=3.1.1
Flask-Limiter~=3.8.0
Flask-JWT-Extended~=4.6.0
gotrue==1.0.3
gunicorn==21.2.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==2.10
importlib-metadata==6.8.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
langcodes==3.3.0
MarkupSafe==2.1.3
murmurhash==1.0.9
numpy==1.25.2
packaging==23.1
pathy==0.10.2
pluggy==1.2.0
postgrest==0.10.8
preshed==3.0.8
psycopg~=3.2
py==1.11.0
pycparser==2.21
pydantic==2.2.1
pydantic_core==2.6.1
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
realtime==1.0.0
regex==2023.8.8
requests==2.31.0
rfc3986==1.5.0
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.7
storage3==0.5.4
StrEnum==0.4.15
supabase==1.0.4
supafunc==0.2.3
thinc==8.1.12
toml==0.10.2
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.7.1
urllib3==1.26.18
wasabi==1.1.2
websockets==10.4
Werkzeug==3.0.1
zipp==3.16.2
pytest-mock~=3.12.0
Authlib~=1.3.1
PyJWT~=2.9.0
marshmallow~=3.22.0